# mcp-server-generator/generated/claim-center-mcp-1a82bf09/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Insurance ClaimCenter API",
    "description": "Report, search, and manage insurance claims.",
    "version": "1.0.0"
  },
  "paths": {
    "/health": {
      "get": {
        "summary": "Health",
        "operationId": "claim_center_health",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object",
                  "title": "Response Claim Center Health"
                }
              }
            }
          }
        }
      }
    },
    "/claims": {
      "get": {
        "summary": "Search Claims",
        "operationId": "search_claims",
        "parameters": [
          {
            "name": "policy_number",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Policy Number"
            }
          },
          {
            "name": "customer_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Customer Id"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "reported",
                    "investigating",
                    "approved",
                    "closed",
                    "denied"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Claim"
                  },
                  "title": "Response Search Claims"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Report Claim",
        "operationId": "report_claim",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportClaimRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Claim"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/claims/{claim_number}": {
      "get": {
        "summary": "Get Claim",
        "operationId": "get_claim",
        "parameters": [
          {
            "name": "claim_number",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Claim Number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Claim"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/claims/{claim_number}/status": {
      "patch": {
        "summary": "Update Claim Status",
        "operationId": "update_claim_status",
        "parameters": [
          {
            "name": "claim_number",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Claim Number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateClaimStatusRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Claim"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Claim": {
        "properties": {
          "claim_number": {
            "type": "string",
            "title": "Claim Number"
          },
          "policy_number": {
            "type": "string",
            "title": "Policy Number"
          },
          "customer_id": {
            "type": "string",
            "title": "Customer Id"
          },
          "loss_date": {
            "type": "string",
            "format": "date",
            "title": "Loss Date"
          },
          "loss_type": {
            "type": "string",
            "enum": [
              "collision",
              "theft",
              "property_damage",
              "water_damage"
            ],
            "title": "Loss Type"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "status": {
            "type": "string",
            "enum": [
              "reported",
              "investigating",
              "approved",
              "closed",
              "denied"
            ],
            "title": "Status"
          },
          "reserve_amount": {
            "type": "number",
            "title": "Reserve Amount"
          },
          "paid_amount": {
            "type": "number",
            "title": "Paid Amount"
          }
        },
        "type": "object",
        "required": [
          "claim_number",
          "policy_number",
          "customer_id",
          "loss_date",
          "loss_type",
          "description",
          "status",
          "reserve_amount",
          "paid_amount"
        ],
        "title": "Claim"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ReportClaimRequest": {
        "properties": {
          "policy_number": {
            "type": "string",
            "maxLength": 30,
            "minLength": 3,
            "title": "Policy Number"
          },
          "customer_id": {
            "type": "string",
            "maxLength": 30,
            "minLength": 3,
            "title": "Customer Id"
          },
          "loss_date": {
            "type": "string",
            "format": "date",
            "title": "Loss Date"
          },
          "loss_type": {
            "type": "string",
            "enum": [
              "collision",
              "theft",
              "property_damage",
              "water_damage"
            ],
            "title": "Loss Type"
          },
          "description": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 10,
            "title": "Description"
          },
          "estimated_loss": {
            "type": "number",
            "maximum": 10000000.0,
            "exclusiveMinimum": 0.0,
            "title": "Estimated Loss"
          }
        },
        "type": "object",
        "required": [
          "policy_number",
          "customer_id",
          "loss_date",
          "loss_type",
          "description",
          "estimated_loss"
        ],
        "title": "ReportClaimRequest"
      },
      "UpdateClaimStatusRequest": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "reported",
              "investigating",
              "approved",
              "closed",
              "denied"
            ],
            "title": "Status"
          },
          "note": {
            "type": "string",
            "maxLength": 500,
            "minLength": 3,
            "title": "Note"
          }
        },
        "type": "object",
        "required": [
          "status",
          "note"
        ],
        "title": "UpdateClaimStatusRequest"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}